pyproject.toml
[tool.poetry]
name = "syt-bot"
version = "0.1.0"
description = "Simple YouTube Discord Bot"
authors = ["E <e@pastel.codes>"]
packages = [
    { include = "syt_bot", from = "src" },
]

[tool.poetry.scripts]
start-syt = 'syt_bot.main:main'

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {extras = ["voice"], version = "^1.7.3"}
python-dotenv = "^0.20.0"
youtube_dl = "^2021.12.17"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]
coverage = {extras = ["toml"], version = "^6.4"}
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pytest = "^7.1.2"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
